Fix size_t to int conversion warnings in Reports.cpp and StaffModule.cpp - Added explicit static_cast to prevent data loss warnings - Fixed padding calculations to properly handle size_t to int conversions

+++ src/StaffModule.cpp
@@ -743,10 +743,10 @@ void StaffModule::displayTableHeader(const string& title) {
     // Highlighted title: Yellow text on Blue background
     ColorUtils::setColor(WHITE);
     cout << "|";
-    int padding = (60 - title.length()) / 2;
+    int padding = static_cast<int>((60 - static_cast<int>(title.length())) / 2);
     for (int i = 0; i < padding; i++) cout << " ";
     ColorUtils::printColoredBG(title, YELLOW, BLUE);
-    for (int i = 0; i < (60 - title.length() - padding); i++) cout << " ";
+    for (int i = 0; i < static_cast<int>(60 - title.length() - padding); i++) cout << " ";
     ColorUtils::setColor(WHITE);
     cout << "|" << endl;
